refactor: consolidate test structure into back/tests/
Fixes #81

## Changes

- Moved app/tests/mocks/ → tests/mocks/
- Moved app/tests/test_physical_controls_integration.py → tests/integration/hardware/
- Moved app/tests/unit/test_state_manager.py → tests/unit/services/test_state_manager_app_legacy.py
- Appended mock fixtures from app/tests/conftest.py to tests/conftest.py
- Deleted app/tests/ directory

## Benefits

- Single, clear location for all tests (back/tests/)
- Follows Python standard practices
- No confusion about where to add new tests
- Fixes metrics workflow that was only running 2 tests instead of 146
