--- back/app/tests/__init__.py
@@ -1,9 +1,0 @@
-# Copyright (c) 2025 Jonathan Piette
-# This file is part of TheOpenMusicBox and is licensed for non-commercial use only.
-# See the LICENSE file for details.
-
-"""Test package for TheOpenMusicBox backend.
-
-This package contains unit tests, integration tests, and mock implementations
-for testing the backend services and controllers.
-"""
