--- back/app/tests/integration/__init__.py
@@ -1,5 +1,0 @@
-# Copyright (c) 2025 Jonathan Piette
-# This file is part of TheOpenMusicBox and is licensed for non-commercial use only.
-# See the LICENSE file for details.
-
-"""Integration tests package for TheOpenMusicBox backend API endpoints."""
