--- back/app/tests/pytest_plugins/__init__.py
@@ -1,1 +1,0 @@
-# Pytest plugins initialization
