+++ back/tests/mocks/mock_nfc_service.py
