+++ back/tests/unit/services/test_state_manager_app_legacy.py
